% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-file.R
\name{acceptable_var_name}
\alias{acceptable_var_name}
\title{Is this an acceptable variable name}
\usage{
acceptable_var_name(text)
}
\arguments{
\item{text}{character vector to test}
}
\value{
logical scalar
}
\description{
Is this an acceptable variable name
}
\keyword{internal}
