% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-file.R
\name{check_equation_input}
\alias{check_equation_input}
\title{Basic equation checks}
\usage{
check_equation_input(equation_vector)
}
\arguments{
\item{equation_vector}{character vector of equations, as written}
}
\value{
TRUE, invisibly; stops if any check fails
}
\description{
Basic equation checks
}
\keyword{internal}
