% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-file.R
\name{check_var_vector}
\alias{check_var_vector}
\title{Check a vector of variable names}
\usage{
check_var_vector(variable_vector, name = "variables")
}
\arguments{
\item{variable_vector}{character vector of names}

\item{name}{what is being checked, for the message}
}
\value{
TRUE, invisibly; stops if any check fails
}
\description{
Check a vector of variable names
}
\keyword{internal}
