% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-file.R
\name{check_variable_conflict}
\alias{check_variable_conflict}
\title{Check that two sets of variables do not overlap}
\usage{
check_variable_conflict(
  var1,
  var2,
  what1 = "the first set",
  what2 = "the second set"
)
}
\arguments{
\item{var1, var2}{character vectors}

\item{what1, what2}{what each vector is, for the message}
}
\value{
TRUE, invisibly; stops on an overlap
}
\description{
Check that two sets of variables do not overlap
}
\keyword{internal}
