% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen-runtime.R
\name{cpp_dbl_array}
\alias{cpp_dbl_array}
\title{Format a double vector as a C array initialiser}
\usage{
cpp_dbl_array(x, per_line = 12L)
}
\description{
Format a double vector as a C array initialiser
}
\keyword{internal}
