% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen-runtime.R
\name{cpp_int_array}
\alias{cpp_int_array}
\title{Format an integer vector as a C array initialiser}
\usage{
cpp_int_array(x, per_line = 20L)
}
\description{
Format an integer vector as a C array initialiser
}
\keyword{internal}
