% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen-expr.R
\name{cpp_tshift}
\alias{cpp_tshift}
\title{Add a lag to a time offset}
\usage{
cpp_tshift(toff, k = 0L, vterms = character(0))
}
\arguments{
\item{toff}{time offset, as built by `cpp_t0()`}

\item{k}{integer number of periods to go back}

\item{vterms}{character vector of C++ expressions to additionally go back by}
}
\description{
Add a lag to a time offset
}
\keyword{internal}
