% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-equations.R
\name{create_equations_list}
\alias{create_equations_list}
\title{Build the equations table}
\usage{
create_equations_list(eq_list, verbose = TRUE)
}
\arguments{
\item{eq_list}{character vector of equations, as written in the model file}

\item{verbose}{report names that had to be altered}
}
\value{
data.frame with `id`, `name`, `equation`, `LHS`, `RHS`, `formula`
}
\description{
Build the equations table
}
\keyword{internal}
