% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-checks.R
\name{data_model_checks}
\alias{data_model_checks}
\title{Are all the model's variables in the database}
\usage{
data_model_checks(model, database, quiet = TRUE)
}
\arguments{
\item{model}{a `thor_model`}

\item{database}{the database to check}

\item{quiet}{logical. FALSE to report success as well as failure.}
}
\value{
TRUE if every model variable is present
}
\description{
Are all the model's variables in the database
}
