% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-equations.R
\name{generate_equations_index}
\alias{generate_equations_index}
\title{Generate equation ids}
\usage{
generate_equations_index(equation_list)
}
\arguments{
\item{equation_list}{character vector of equations}
}
\value{
character vector of `eq_###` ids, zero-padded to a fixed width
}
\description{
Generate equation ids
}
\keyword{internal}
