% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-file.R
\name{is_in_formulas}
\alias{is_in_formulas}
\title{Which of these variables actually occur in the equations}
\usage{
is_in_formulas(variables_to_test, formulas, print_type = "", verbose = TRUE)
}
\arguments{
\item{variables_to_test}{character vector of names}

\item{formulas}{character vector of equations}

\item{print_type}{what is being tested, for the message}

\item{verbose}{report what was found as well as what was dropped}
}
\value{
the subset of `variables_to_test` that occurs in `formulas`
}
\description{
Which of these variables actually occur in the equations
}
\keyword{internal}
