% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{is_r_backend}
\alias{is_r_backend}
\title{Is this an R (not compiled) backend}
\usage{
is_r_backend(backend)
}
\arguments{
\item{backend}{character scalar}
}
\description{
Is this an R (not compiled) backend
}
\keyword{internal}
