% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolchain.R
\name{makevars_in_effect}
\alias{makevars_in_effect}
\title{Makevars files R applies to compilations, in the order R reads them}
\usage{
makevars_in_effect(makevars_user = Sys.getenv("R_MAKEVARS_USER", unset = NA))
}
\arguments{
\item{makevars_user}{value of `R_MAKEVARS_USER` to assume, NA for unset}
}
\value{
a named character vector of existing files (`site`, `user`)
}
\description{
Makevars files R applies to compilations, in the order R reads them
}
\keyword{internal}
