% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-equations.R
\name{model_max_lag}
\alias{model_max_lag}
\title{The largest lag any of a model's formulas reaches}
\usage{
model_max_lag(formulas)
}
\arguments{
\item{formulas}{character vector of `new_formula`s}
}
\value{
list(k, variable)
}
\description{
The largest lag any of a model's formulas reaches
}
\keyword{internal}
