% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{model_source_example}
\alias{model_source_example}
\title{Path to a sample model file}
\usage{
model_source_example(path_only = FALSE)
}
\arguments{
\item{path_only}{TRUE to return the path instead of opening the file}
}
\value{
the path to the example model
}
\description{
Path to a sample model file
}
\examples{
model_source_example(TRUE)
}
