% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-checks.R
\name{na_report_variables_times}
\alias{na_report_variables_times}
\title{Which variables have missing values over given periods}
\usage{
na_report_variables_times(database, times, variables, index_time = "date")
}
\arguments{
\item{database}{the database to check}

\item{times}{periods to test, as found in `index_time`}

\item{variables}{variables to check}

\item{index_time}{name of the time column}
}
\value{
character vector of the variables with an NA over `times`
}
\description{
Which variables have missing values over given periods
}
