% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{newdiff}
\alias{newdiff}
\alias{delta}
\title{Differences that keep the vector's length}
\usage{
newdiff(n = 1, x)

delta(n = 1, x)
}
\arguments{
\item{n}{number of periods to difference over}

\item{x}{numeric vector}
}
\value{
numeric vector the same length as `x`, with `n` leading NAs
}
\description{
Differences that keep the vector's length
}
