% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-file.R
\name{parser_lag_delta_check}
\alias{parser_lag_delta_check}
\title{Can the lags and deltas be parsed safely}
\usage{
parser_lag_delta_check(formula_list)
}
\arguments{
\item{formula_list}{character vector of equations}
}
\value{
logical scalar
}
\description{
Can the lags and deltas be parsed safely
}
\keyword{internal}
