% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian-sparse.R
\name{print.thor_sparse_jacobian}
\alias{print.thor_sparse_jacobian}
\title{Print a sparse symbolic jacobian}
\usage{
\method{print}{thor_sparse_jacobian}(x, ...)
}
\arguments{
\item{x}{a `thor_sparse_jacobian`}

\item{...}{ignored}
}
\value{
`x`, invisibly
}
\description{
Print a sparse symbolic jacobian
}
