% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{print_build_timings}
\alias{print_build_timings}
\title{Print the durations of a build on one line}
\usage{
print_build_timings(x)
}
\arguments{
\item{x}{as returned by [build_timings()]}
}
\value{
`x`, invisibly
}
\description{
Print the durations of a build on one line
}
\keyword{internal}
