% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian-sparse.R
\name{strip_delta}
\alias{strip_delta}
\title{Replace `delta(n, x)` by `x` throughout an expression}
\usage{
strip_delta(e)
}
\arguments{
\item{e}{a language object}
}
\value{
a language object
}
\description{
Only valid for differentiating with respect to a current-period variable;
see the note above.
}
\keyword{internal}
