% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-checks.R
\name{time_model_checks}
\alias{time_model_checks}
\title{Is this column usable as a time index}
\usage{
time_model_checks(database, index_time)
}
\arguments{
\item{database}{the database to check}

\item{index_time}{name of the column}
}
\value{
TRUE if the column can be used as a time index
}
\description{
It must have no duplicates, no NAs, and be sorted in ascending order.
}
