% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolchain.R
\name{toolchain_findings}
\alias{toolchain_findings}
\title{What can be said about the toolchain without compiling anything}
\usage{
toolchain_findings(makevars_user = Sys.getenv("R_MAKEVARS_USER", unset = NA))
}
\arguments{
\item{makevars_user}{value of `R_MAKEVARS_USER` to assume}
}
\value{
a list: `makevars` (files in effect), `problems` (character)
}
\description{
What can be said about the toolchain without compiling anything
}
\keyword{internal}
