% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen-expr.R
\name{var_col}
\alias{var_col}
\title{Column of one variable}
\usage{
var_col(vidx, name)
}
\arguments{
\item{vidx}{a map from [var_index()]; a plain named vector also works}

\item{name}{variable name}
}
\value{
the column number, or NULL if the variable is unknown
}
\description{
Column of one variable
}
\keyword{internal}
