% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-r.R
\name{without_jit}
\alias{without_jit}
\title{Switch R's byte-code compiler off until the calling function returns}
\usage{
without_jit(envir = parent.frame())
}
\arguments{
\item{envir}{the frame whose exit restores the previous setting}
}
\description{
See the note at the top of this file.
}
\keyword{internal}
