% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolchain.R
\name{withr_dir}
\alias{withr_dir}
\title{Evaluate an expression with a different working directory}
\usage{
withr_dir(dir, code)
}
\arguments{
\item{dir}{directory}

\item{code}{expression}
}
\description{
Evaluate an expression with a different working directory
}
\keyword{internal}
